Move content script injection to manifest for speed improvements. Remove blacklistedController, move logic into blockController and contentController. Since content scripts are now always included, apply blacklist logic to each controller to make sure they only run when not blocked.

+++ js/controllers/content/contentController.js
@@ -3,30 +3,39 @@ var terafm = window.terafm;
 (function(db, options, recoveryDialog, initHandler, Events, editableManager) {
 	'use strict';
 
-	// Load extension options into memory
-	options.loadFromChromeStorage(function() {
+	terafm.blacklist.isBlocked(window.location.hostname, function(blocked) {
+		if(blocked) {
+			terafm.isBlocked = true;
 		
-		// Initiate DB, populate in-memory storage
-		db.init(function() {
+		} else {
+			
+			// Load extension options into memory
+			options.loadFromChromeStorage(function() {
+				
+				// Initiate DB, populate in-memory storage
+				db.init(function() {
 
-			// Run init handlers
-			initHandler.executeInitHandlers();
-		});
+					// Run init handlers
+					initHandler.executeInitHandlers();
+					console.log('Typio is ready!')
+				});
+			});
+		}
 	});
 
 
 	// Messages from background.js
 	chrome.runtime.onMessage.addListener(function(request, sender, sendResponse) {
-
-		// Used to check if content scripts are already injected
-		if(request.action === 'ping') {
-			sendResponse(true);
-		
-		} else if(request.action === 'clearData') {
+		if(request.action === 'clearData') {
+			if(terafm.isBlocked) return terafm.blockController.warn();
 			db.deleteAllDataForDomain();
 			terafm.toastController.create('Database cleared for ' + window.location.hostname);
 		}
 	});
+
+
+
+
 	
 
 })(terafm.db, terafm.options, terafm.recoveryDialog, terafm.initHandler, terafm.Events, terafm.editableManager);
